Make the source ASCII only.
AppleScript uses Mac-Roman or UTF-16 (with a BOM); at least, that's
what AppleScript Editor saves in .applescript files.  (For what it's
worth, .scpt files have text objects in both Mac-Roman and UTF-16,
with the Mac-Roman version using ? for unencodable characers.)  Since
git doesn't have a mechanism to specify a file's encoding (it sees
file contents as binary blobs, except that it'll sometimes edit the
line endings), it needs to be saved in one of those formats if we want
a checkout from git to work as a .applescript file.  But, github will
display it as ISO-8859-1.  (The raw view sets the Content-Type, and
the web view will translate it from to UTF-8 and send it to the
browser that way.)  If a user does a copy/paste from the browser into
AppleScript Editor (like my blog says to do), then they'll get a
different character.  To allow either a git clone or a copy/paste from
github to work, we have to use ASCII.