Add user-configurable callbacks.
One of these is shouldNotify, and this commit also adds ignoreBuddies
to its capabilities.

This requires a change to the "on growl" signature, since I want the
user hooks to have some consistency in some of the parameter names,
but also want the hooks to not need iChat's terms.  This means that
the "status" parameter was changed to "showingStatus".