Update to Growl development team's recommendations.
This will check to see if Growl is running before sending
notifications, and it will look for Growl by bundle id instead of by
name.  (The latter removes the problem with the different application
name between Growl 1.2 and 1.3.)

Thanks to Chris Forsythe (the lead developer of Growl) for the
information.  This is based on his commit on
https://github.com/ChrisForsythe/iChat-Growl
7a974c2505efd0c0603a14c6b13ed1c9548e3d34, but was edited to fit in
with the other changes at the same time.